refactor: MainPage 컴포넌트를 모듈화하여 코드 분리
- ThemeCategories 컴포넌트를 분리하여 테마 카테고리 섹션을 구성
- TrendingGifts 컴포넌트를 분리하여 실시간 급상승 선물 랭킹 섹션을 구성
- MainPage 컴포넌트에서 분리된 컴포넌트들을 사용하여 전체 레이아웃 구성
